requirements.txt
# django==1.11
Pillow==4.3.0
requests==2.18.4
mysqlclient==1.3.12
django-el-pagination==3.1.0
django-tinymce==2.7.0
django-tagulous==0.12.0
bs4